Рефакторинг - вынесение кнопок с Favorite. 1. Вынесение кнопок переключения таймфреймов на графике в отдельный компонент.
