+++ pages/Favorite/Favorite.styles.js
@@ -57,51 +57,14 @@ export const styles = StyleSheet.create({
     paddingBottom: "10"
   },
 
-  chartButtons: {
-    flexDirection: "row",
-    gap: 10,
-    backgroundColor: "rgba(50, 48, 49, 0.8)",
-    borderWidth: 0.5,
-    borderColor: "wheat",
-    padding: 7,
-    borderRadius: 15,
-
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 2,
-      height: 4
-    },
-    shadowOpacity: 0.15,
-    shadowRadius: 6,
-    elevation: 5 // Для Android
-  },
-  chartButton: {
-    marginTop: 0,
-    backgroundColor: "darkgray",
-    borderRadius: 5,
-    paddingVertical: 7,
-    paddingHorizontal: 20,
-    alignItems: "center"
-  },
-  activeButton: {
-    backgroundColor: "#000"
-  },
-  activeButtonText: {
-    color: "#fff"
-  },
-  buttonText: {
-    color: "#000",
-    fontSize: RFValue(10),
-    fontWeight: "bold"
-  },
   text: { color: "white", paddingBottom: 5, fontSize: RFValue(12) },
-  text1: { color: "white", fontSize: RFValue(12) },
+  text1: { color: "white", paddingTop: 5, fontSize: RFValue(12) },
   textZ: {
     color: "wheat",
     fontSize: RFValue(12)
   },
   chartButtonsClose: {
-    marginTop: 5,
+    marginTop: 10,
     flexDirection: "row",
     gap: 10,
     backgroundColor: "rgba(50, 48, 49, 0.8)",
